requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==2.2
django-appconf==1.0.3
django-ckeditor==5.8.0
django-debug-toolbar==2.2
django-imagekit==4.0.2
django-js-asset==1.2.2
django-location-field==2.1.0
django-modeltranslation==0.14.4
idna==2.8
pilkit==2.0
Pillow==6.2.1
psycopg2==2.8.4
python-decouple==3.1
python-gettext==4.0
pytz==2019.3
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
